Start rewriting the config parser with Rhai
Starlark is neat, but its rust implementation suffers from the nightly
dependency: Nightly moves too fast and too divergently with
gazebo/starlark for me to keep up with the various incompatibilities
that are introduced in the various large-scale rewrites of the
different crates/components involved.

I want this to run on stable, and Rhai offers me that, while giving me
a very similar language and feature set to starlark, with a somewhat
nicer rust embedding experience? Sign me up.